Extract code about legacy sources
Preparing that similar code will be used from the main and the renderer
process.
